Primitive관련 클래스의 이름을 Rasterizer로 변경. Line 및 Point도 래스터 방식 변경.

+++ SoftwareRenderer/Renderer/Vertex.h
@@ -19,7 +19,7 @@ namespace Renderer
 		Vector4 color;
 		Vector3 clip;
 		Vector2Int screen;
-		int depth;
+		float depth;
 	};
 
 }
